+++ src/badge/Badge.unit.tsx
@@ -5,7 +5,7 @@ import { CheckIcon as StyledCheckIcon } from '../icon/checkIcon'
 import { Badge } from './Badge'
 
 it('Should read the aria-label attribute and not the children if both are provided', () => {
-  const wrapper = shallow(<Badge ariaLabel="Unread messages">12</Badge>)
+  const wrapper = shallow(<Badge aria-label="Unread messages">12</Badge>)
   expect(wrapper.find('span > span').prop('aria-hidden')).toBe(true)
 })