+++ src/text/Text.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 import cc from 'classcat'
 
 import { replaceNewLineWithBR } from '../_utils'
+import { A11yProps, pickA11yProps } from '../_utils/interfaces'
 
 export enum TextDisplayType {
   BODY = 'body',
@@ -22,32 +23,30 @@ export enum TextTagType {
   SPAN = 'span',
 }
 
-export interface TextProps {
+export interface TextProps extends A11yProps {
   readonly className?: string
   readonly children: string | number | React.ReactNode
   readonly display?: TextDisplayType
   readonly tag?: TextTagType
   readonly textColor?: string
   readonly newlineToBr?: boolean
-  readonly role?: string
-  readonly ariaLabel?: string
   readonly itemProp?: string
 }
 
 const baseClassName = 'kirk-text'
 const cssColorRegex = /^#([0-9a-f]{3}|[0-9a-f]{6})$/i
 
-export const Text = ({
-  className,
-  children,
-  display = TextDisplayType.BODY,
-  tag = TextTagType.SPAN,
-  textColor,
-  newlineToBr = true,
-  role,
-  ariaLabel,
-  itemProp = null,
-}: TextProps) => {
+export const Text = (props: TextProps) => {
+  const {
+    className,
+    children,
+    display = TextDisplayType.BODY,
+    tag = TextTagType.SPAN,
+    textColor,
+    newlineToBr = true,
+    itemProp = null,
+  } = props
+  const a11yAttrs = pickA11yProps<TextProps>(props)
   const displayClassName = `${baseClassName}-${display}`
   const Tag = tag
 
@@ -55,11 +54,10 @@ export const Text = ({
 
   return (
     <Tag
-      role={role}
-      aria-label={ariaLabel}
       className={cc([baseClassName, displayClassName, className])}
       {...inlineStyle}
       {...(itemProp && { itemProp })}
+      {...a11yAttrs}
     >
       {typeof children === 'string' && newlineToBr ? replaceNewLineWithBR(children) : children}
     </Tag>