+++ src/tripCard/story.tsx
@@ -11,7 +11,7 @@ const stories = storiesOf('Widgets|TripCard', module)
 stories.addDecorator(withKnobs)
 
 const tripCardConfig = () => ({
-  ariaLabel: text(
+  'aria-label': text(
     'aria-label',
     'Pick-up point: Paris, Drop-off point: Bordeaux, Departure time: 09:00, warning',
   ),