+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 [...]
 
+# v38.0.0 (24/07/2020)
+
+- **[BREAKING CHANGE]** Removed all `ariaX` props from components and switched to support all common `aria-x` attributes with `A11yProps` interface.
+
+# v37.2.0 (23/07/2020)
+
+- **[UPDATE]** Add `noHorizontalSpacing` prop to disable horizontal padding on `Modal` component.
+- **[NEW]** Added `FilterBar` component
+
 # v37.1.1 (22/07/2020)
 
 - **[FIX]** Missing props exports in index file for `ItemAction`, `ItemBigData`, `ItemData`, `ItemEditableInfo`